requirements.txt
deepgram-sdk==v3.10.1
Flask==3.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
transformers==4.41.2
langchain==0.2.9
langchain-community==0.2.5
langchain-core==0.2.43
langchain-huggingface==0.0.3
langchain-openai==0.1.9
langchain-text-splitters==0.2.4
chroma-hnswlib==0.7.3
chromadb==0.5.3
pandas==1.5.3
pydantic==2.10.5  # Ensure to match the version if langchain requires a specific one
librosa == 0.10.2
soundfile == 0.12.1
gradio == 1.0.2
datasets == 2.20.
PyMySQL==1.1.1
